+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- experimental generator to integrate the oscal compliance template
+
 ## [0.6.0] - 2022-06-07
 
 - include USWDS 3.0 for new apps